initial pmm and gdt work

+++ libk/stdlib.c
@@ -1,9 +1,9 @@
 #include "stdlib.h"
 #include <stdbool.h>
 
-void itoa(size_t x, char *str) {
+void itoa(int64_t x, char *str) {
     uint8_t i = 0;
-    size_t r = 0;
+    int64_t r = 0;
     int8_t mult = 1;
     if(x < 0) {
         str[0] = '-';
