+++ libk/stdlib.h
@@ -2,5 +2,5 @@
 #define STDLIB_H
 #include <stddef.h>
 #include <stdint.h>
-void itoa(size_t x, char *str);
+void itoa(int64_t x, char *str);
 #endif
